More refactoring, to remove JKSet dependences
